Icon and logo w/ Bitmgothic & Dark Oak Planks
The pixel art mug was made by me (though the shape of the foam is the
same as the grass block's side texture from Minecraft).

The wooden sign is made by me from the dark oak planks texture from
Minecraft.

The font used on the sign is Bitmgothic by Nomi
https://www.thenomi.org/